refactor: merge timestats package with metrics package
