+++ benchttp/metrics/timestats_sort.go
@@ -1,4 +1,4 @@
-package timestats
+package metrics
 
 import (
 	"time"
